Added basic save function
How to save and load: you have 16 save slots available. You change the active save slot by pressing the encoder. Currently there is no way to track which save slot is currently active. I'm working on that. Then if you turn the encoder to the right, you will load the patch saved in that slot. Turning the encoder to the left to save (and overwrite) the patch. Beware, if you try to load an empty save slot, the Daisy will crash and you will need to reboot it.